+++ src/test/java/org/scify/jedai/configuration/OptimizeDirtyMoviesDataset.java
@@ -51,8 +51,8 @@ public class OptimizeDirtyMoviesDataset {
     private final static int NO_OF_TRIALS = 100;
 
     static float getTotalComparisons(List<AbstractBlock> blocks) {
-        int originalComparisons = 0;
-        originalComparisons = blocks.stream().map(AbstractBlock::getNoOfComparisons).reduce(originalComparisons, Integer::sum);
+        long originalComparisons = 0;
+        originalComparisons = blocks.stream().map(AbstractBlock::getNoOfComparisons).reduce(originalComparisons, Long::sum);
         System.out.println("Original comparisons\t:\t" + originalComparisons);
         return originalComparisons;
     }
